python-orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wordsynk-orchestrator"
version = "0.1.0"
dependencies = [
    "tomlkit>=0.11",
    "pydantic>=2.0",
    "frida-tools>=12.0",
    "pure-python-adb>=0.6",
    "python-telegram-bot>=20.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "black", "ruff"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
